requirements.txt
Flask
flask-cors
fpdf
pandas
odfpy
PyMuPDF
transformers
torch
sentence-transformers
